04.06.2018 changes in labels. Letter title + subject changed to: פריטים מושאלים שיש להחזירם לספרייה.
added template: "senderReceiverExtended"
changed books table
added library details at the footer.
added donotreply template at the footer

To Do:
need to understand if the library details are actually retreived
The template resposible for that: "senderReceiverExtended", from some reason the details are not retrieved. Maybe because letters are not current - need to test during this week. 04-07.06.18

Need to understand where is the reason for sending the letter is located.

Send question to basecamp regarding the "due_date_shortened_reason" xml element, is that useful?

10.06.2018 == shortened date reason on review with exlibris